+++ Scripts/Bash/reissueFileVaultKey/CHANGELOG.md
@@ -1,9 +1,14 @@
 # Reissue FileVault Key release history
 
+## 3.0.0 - 2023-02-23
+
+* Changed `LAST_SUPPORTED_OS_VERSION` to `SUPPORTED_OS_MAJOR_VERSION` [`Breaking`].
+* Minor code tweaks
+
 ## 2.0.0 - 2023-02-23
 
 * Added better notifications and alerts.
-* Changed `displayDialog` icon parameter to allow icon resource names or ids
+* Changed `displayDialog` icon parameter to allow icon resource names or IDs.
 * Changed `displayDialog` parameters position to make more sense.
 * Changed last supported OS version to macOS 12.